Split user info class into UserInfo and UserInfoPublic inheriting an abstract class
