+++ app/src/main/java/com/group4sweng/scranplan/UserInfoPublic.java
@@ -1,0 +1,14 @@
+package com.group4sweng.scranplan;
+
+import java.io.Serializable;
+
+public class UserInfoPublic extends UserBase implements Serializable {
+
+    enum Privacy {
+        PUBLIC,
+        RESTRICTED,
+        PRIVATE
+    }
+
+
+}
